rename test_completeness.py to test_redundancy.py
