+++ tests/stats/test_redundancy.py
